refactor: add lambda to split cur files into micro files
