+++ modules/integrations/splunk_aws_billing/role.tf
@@ -22,6 +22,16 @@ data "aws_iam_policy_document" "lambda_policy_document" {
 
   statement {
     actions = [
+      "s3:PutObject",
+    ]
+    effect = "Allow"
+    resources = [
+      "arn:aws:s3:::${aws_s3_bucket.aws_billing_report.id}/tmp/*"
+    ]
+  }
+
+  statement {
+    actions = [
       "secretsmanager:GetSecretValue"
     ]
     effect = "Allow"
